Switch to parcel, reorganize the project
