+++ custom.d.ts
@@ -1,0 +1,9 @@
+declare module '*.jpg' {
+  const content: any;
+  export default content;
+}
+
+declare module '*.scss' {
+  const content: any;
+  export default content;
+}
