+++ site/components/pages/Docs/chapters/index.tsx
@@ -1,0 +1,3 @@
+export * from './Buttons';
+export * from './Cards';
+export * from './Usage';
