+++ site/components/shared/Demo.scss
@@ -1,0 +1,7 @@
+.Demo {
+  margin-bottom: 1rem;
+
+  pre code {
+    background-color: white;
+  }
+}
