+++ site/components/shared/Highlight.tsx
@@ -2,7 +2,7 @@ import * as React from 'react';
 import SyntaxHighlighter from 'react-syntax-highlighter';
 import { SyntaxHighlighterProps } from 'react-syntax-highlighter/light';
 import { github } from 'react-syntax-highlighter/styles/hljs';
-import './Highlight.css';
+import './Highlight.scss';
 
 export function Highlight(props: SyntaxHighlighterProps) {
   return (
