+++ site/components/shared/index.tsx
@@ -1,0 +1,2 @@
+export * from './Demo';
+export * from './Highlight';
