+++ site/index.html
@@ -13,5 +13,6 @@
   You need to enable JavaScript.
 </noscript>
 <div id="root"></div>
+<script src="./index.tsx"></script>
 </body>
 </html>
