+++ site/index.scss
@@ -1,3 +1,5 @@
+@import '../scss/matilda.scss';
+
 html,
 body,
 #root {
