--- site/package.json
@@ -1,42 +1,0 @@
-{
-  "name": "site",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "classnames": "^2.2.6",
-    "history": "^4.7.2",
-    "mobx": "^5.0.3",
-    "mobx-react": "^5.2.3",
-    "mobx-react-router": "^4.0.4",
-    "pretty": "^2.0.0",
-    "react": "^16.4.1",
-    "react-dom": "^16.4.1",
-    "react-router-dom": "^4.3.1",
-    "react-scripts-ts": "2.16.0",
-    "react-syntax-highlighter": "^8.0.1",
-    "reactstrap": "^6.3.0"
-  },
-  "scripts": {
-    "start": "react-scripts-ts start",
-    "build": "react-scripts-ts build",
-    "test": "react-scripts-ts test --env=jsdom",
-    "eject": "react-scripts-ts eject",
-    "lint": "tslint 'src/**/*.{ts,tsx}' --fix"
-  },
-  "devDependencies": {
-    "@types/classnames": "^2.2.5",
-    "@types/history": "^4.6.2",
-    "@types/jest": "^23.1.6",
-    "@types/node": "^10.5.2",
-    "@types/react": "^16.4.6",
-    "@types/react-dom": "^16.0.6",
-    "@types/react-router-dom": "^4.2.7",
-    "@types/react-syntax-highlighter": "0.0.5",
-    "@types/reactstrap": "^6.0.0",
-    "prettier": "^1.13.7",
-    "typescript": "^2.9.2"
-  },
-  "prettier": {
-    "singleQuote": true
-  }
-}
