--- site/src/components/pages/Docs/chapters/index.tsx
@@ -1,2 +1,0 @@
-export * from './Usage';
-export * from './Buttons';
