--- site/src/components/shared/Demo.css
@@ -1,7 +1,0 @@
-.Demo {
-  margin-bottom: 1rem;
-}
-
-.Demo pre code {
-  background-color: white;
-}
