--- site/src/components/shared/index.tsx
@@ -1,1 +1,0 @@
-export * from './Demo';
